Move command runner to @actions/exec
