+++ packages/exec/package.json
@@ -36,6 +36,7 @@
     "url": "https://github.com/actions/toolkit/issues"
   },
   "dependencies": {
+    "@actions/core": "^1.10.1",
     "@actions/io": "^1.0.1"
   }
 }
